질문 삭제 및 비로그인 사용자 질문등록 제한

+++ app/src/main/java/com/example/qp/QuestionResponse.kt
@@ -42,7 +42,7 @@ data class QuestionPost(
 )
 
 data class UserInfo(
-    @SerializedName(value="userId")val userId: Long,
+    @SerializedName(value="userId")val userId: Int,
     @SerializedName(value="profileImage")val profileImage: String,
     @SerializedName(value="role")val role: String
 ):Serializable
